Update backEnd.js
I added a new system to understand when the user tried to modify a key by pressing a computer key and not a piano key and react differently. Here, when you press a computer key to remove it, so you press 'Escape' just after, now the code will understand that you want to remove all the piano keys associated to this computer key and remove all of the assignations of this key.